Move components into data folder
