+++ src/components/data/DataPlotter.vue
@@ -1,51 +1,39 @@
-<template id="data-table">
-  <div class="data-table">
-    <v-table v-show="title !== ''">
-      <template v-slot:default>
-        <thead>
-          <tr>
-            <th class="text-center" v-html="$t('table.time')" />
-            <th class="text-center" v-html="title" />
-          </tr>
-        </thead>
-      </template>
-    </v-table>
-
+<template id="data-plotter">
+  <div class="data-plotter">
     <v-alert v-show="alert.value" type="warning" v-html="alert.msg" />
 
-    <v-card id="scroll-target" class="overflow-y-auto" max-height="500">
-      <v-table v-scroll:#scroll-target="onScroll">
-        <template v-slot:default>
-          <tbody>
-            <tr v-for="(date, index) in data.time" :key="index">
-              <td v-html="formatDate(date)" />
-              <td v-html="data.value[index]" />
-            </tr>
-          </tbody>
-        </template>
-      </v-table>
+    <v-card min-height="600px">
+      <div :style="{ visibility: loading ? 'visible' : 'hidden' }">
+        <v-progress-linear striped indeterminate color="primary" />
+      </div>
+      <div :style="{ visibility: !loading ? 'visible' : 'hidden' }">
+        <v-container>
+          <v-row justify="center" align="end">
+            <div :id="'plotly-chart-' + choices_.collection.id" />
+          </v-row>
+        </v-container>
+      </div>
     </v-card>
   </div>
 </template>
 
 <script>
+import Plotly from "plotly.js-dist-min";
 import { defineComponent } from "vue";
+import { mdiDownload } from "@mdi/js";
 
 let oapi = window.VUE_APP_OAPI;
 
 export default defineComponent({
-  name: "DataTable",
-  template: "#data-table",
+  name: "DataPlotter",
+  template: "#data-plotter",
   props: ["choices", "alert"],
   watch: {
-    choices: {
+    choices_: {
       handler(newValue) {
-        if (this.loading === true) {
-          return;
-        }
-        this.loading = true;
         if (newValue.collection !== "" && newValue.datastream !== "") {
-          this.data = {};
+          this.data = [];
+          this.config.modeBarButtonsToAdd = [];
           for (var station of this.choices_.station) {
             this.loadCollection(newValue.collection, station);
           }
@@ -57,24 +45,44 @@ export default defineComponent({
   },
   data: function () {
     return {
+      trace: {
+        type: "scatter",
+        mode: "lines+markers",
+        x: [],
+        y: [],
+      },
       choices_: this.choices,
-      data: {},
+      data: [],
       loading: false,
-      title: "",
-      headerOverflow: 0,
+      layout: {
+        title: "",
+        xaxis: {
+          autorange: true,
+          type: "date",
+          range: [null, null],
+          rangeslider: { range: [null, null] },
+        },
+        yaxis: {
+          type: "linear",
+          autorange: true,
+          title: null,
+        },
+      },
+      config: {
+        modeBarButtonsToAdd: [],
+      },
+      font: { size: 14 },
       alert_: this.alert,
     };
   },
   methods: {
-    onScroll(e) {
-      this.headerOverflow = e.target.scrollTop;
-    },
-    formatDate(date) {
-      return new Date(date).toLocaleString(this.$t("code"), {
-        timeZoneName: "short",
-        timeZone: "UTC",
-        hour12: false,
-      });
+    plot() {
+      var plot = document.getElementById(
+        "plotly-chart-" + this.choices.collection.id
+      );
+      Plotly.purge(plot);
+      Plotly.newPlot(plot, this.data, this.layout, this.config);
+      this.loading = false;
     },
     getCol(features, key) {
       if (key.includes(".")) {
@@ -94,18 +102,23 @@ export default defineComponent({
         });
       }
     },
-    newTable(features, colName) {
-      this.data.time = this.getCol(features, "phenomenonTime");
-      this.data.value = this.getCol(features, colName);
-      this.loading = false;
+    newTrace(features, x, y, station_id) {
+      const Trace = JSON.parse(JSON.stringify(this.trace));
+      Trace.x = this.getCol(features, x);
+      Trace.y = this.getCol(features, y);
+      Trace.name = station_id;
+      this.data.push(Trace);
     },
     async loadCollection(collection, station_id) {
+      this.loading = true;
+      var self = this;
+      const range = this.layout.xaxis.range;
       const title = collection.description;
       this.alert_.msg =
         station_id + this.$t("messages.no_observations_in_collection") + title;
-
-      this.loading = true;
-      var self = this;
+      this.layout.title = title;
+      this.layout.xaxis.range = range;
+      this.layout.xaxis.rangeslider = { range: range };
 
       await this.$http({
         method: "get",
@@ -157,15 +170,30 @@ export default defineComponent({
         })
           .then(function (response) {
             // handle success
-            self.title =
+            self.config.modeBarButtonsToAdd.push({
+              name: "Data Source " + station_id,
+              icon: {
+                width: 24,
+                height: 24,
+                path: mdiDownload,
+              },
+              click: function () {
+                window.location.href = response.request.responseURL;
+              },
+            });
+            var title =
               self.choices_.datastream.name +
               " (" +
               self.choices_.datastream.units +
               ")";
-            self.newTable(
+            self.layout.yaxis.title = title;
+            self.newTrace(
               response.data.features,
-              "observations." + self.choices_.datastream.id + ".value"
+              "phenomenonTime",
+              "observations." + self.choices_.datastream.id + ".value",
+              station_id
             );
+            self.plot();
           })
           .catch(function (error) {
             // handle error
